refactor em sala
