+++ app/src/main/java/br/ufrpe/bsi/mpoo/mpooapp/upkeepxpteam/upkeepxp/equipamento/ui/BuscarEquipamentoActivity.java
@@ -23,7 +23,6 @@ import java.util.ArrayList;
 
 import br.ufrpe.bsi.mpoo.mpooapp.upkeepxpteam.upkeepxp.atividadediaria.ui.AtividadeDiariaActivity;
 import br.ufrpe.bsi.mpoo.mpooapp.upkeepxpteam.upkeepxp.equipamento.dominio.Equipamento;
-import br.ufrpe.bsi.mpoo.mpooapp.upkeepxpteam.upkeepxp.equipamento.negocios.EquipamentoAdapter;
 import br.ufrpe.bsi.mpoo.mpooapp.upkeepxpteam.upkeepxp.equipamento.persistencia.EquipamentoDAO;
 import br.ufrpe.bsi.mpoo.mpooapp.upkeepxpteam.upkeepxp.equipes.ui.EquipesActivity;
 import br.ufrpe.bsi.mpoo.mpooapp.upkeepxpteam.upkeepxp.usuario.ui.CadastraUsuarioActivity;
@@ -87,6 +86,7 @@ public class BuscarEquipamentoActivity extends AppCompatActivity implements Navi
                 //chamar a intent que vai habilitar o cadastro no BD
                 Intent it = new Intent(BuscarEquipamentoActivity.this, CadastrarEquipamentos.class);
                 startActivity(it);
+                finish();
             }
         });
