+++ app/src/main/java/br/ufrpe/bsi/mpoo/mpooapp/upkeepxpteam/upkeepxp/equipamento/ui/EquipamentoAdapter.java
@@ -1,4 +1,4 @@
-package br.ufrpe.bsi.mpoo.mpooapp.upkeepxpteam.upkeepxp.equipamento.negocios;
+package br.ufrpe.bsi.mpoo.mpooapp.upkeepxpteam.upkeepxp.equipamento.ui;
 
 import android.content.Context;
 import android.support.annotation.NonNull;
